refactor: ProofQueryClient
